Refactor and simplify association linking logic
