+++ doma-processor/src/main/java/org/seasar/doma/internal/apt/generator/DaoImplQueryMethodGenerator.java
@@ -45,9 +45,9 @@ import org.seasar.doma.internal.apt.cttype.OptionalIntCtType;
 import org.seasar.doma.internal.apt.cttype.OptionalLongCtType;
 import org.seasar.doma.internal.apt.cttype.SimpleCtTypeVisitor;
 import org.seasar.doma.internal.apt.cttype.StreamCtType;
+import org.seasar.doma.internal.apt.meta.aggregate.AggregateStrategyMeta;
+import org.seasar.doma.internal.apt.meta.aggregate.AssociationLinkerMeta;
 import org.seasar.doma.internal.apt.meta.dao.DaoMeta;
-import org.seasar.doma.internal.apt.meta.entity.AggregateStrategyMeta;
-import org.seasar.doma.internal.apt.meta.entity.AssociationLinkerMeta;
 import org.seasar.doma.internal.apt.meta.parameter.BasicInOutParameterMeta;
 import org.seasar.doma.internal.apt.meta.parameter.BasicInParameterMeta;
 import org.seasar.doma.internal.apt.meta.parameter.BasicListParameterMeta;
