+++ doma-processor/src/main/java/org/seasar/doma/internal/apt/meta/aggregate/AggregateStrategyMeta.java
@@ -13,15 +13,25 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.seasar.doma.internal.apt.meta.entity;
+package org.seasar.doma.internal.apt.meta.aggregate;
 
+import java.util.List;
 import java.util.Objects;
 import org.seasar.doma.internal.apt.cttype.EntityCtType;
+import org.seasar.doma.internal.apt.meta.TypeElementMeta;
 
-public record BiFunctionMeta(EntityCtType source, EntityCtType target, EntityCtType result) {
-  public BiFunctionMeta {
-    Objects.requireNonNull(source);
-    Objects.requireNonNull(target);
-    Objects.requireNonNull(result);
+public record AggregateStrategyMeta(
+    EntityCtType root, String tableAlias, List<AssociationLinkerMeta> associationLinkerMetas)
+    implements TypeElementMeta {
+
+  public AggregateStrategyMeta {
+    Objects.requireNonNull(root);
+    Objects.requireNonNull(tableAlias);
+    Objects.requireNonNull(associationLinkerMetas);
+  }
+
+  @Override
+  public boolean isError() {
+    return false;
   }
 }
